Intercept local links
We want to display a content area that slides over from the right for
all the local links, so we need to use JavaScript for loading those.
Some links are external links (like the ones to amazon.jobs and
amazon.racing), so for any links which have an explicit protocol we just
open the link.